feat: connection query cleanup part 2 - split into core and full/enrichment implementation
#3756

Changes Made
1. Created ConnectionQueryCore.cs — Contains:
• HasConnection methods (both overloads)
• GetBaseConnectionsAsync — new public entry point returning List<ConnectionQueryBaseRecord> without enrichment
• BuildBaseQueryFromOthers / BuildBaseQueryToOthers (now protected)
• LoadPackagesByKeyAsync / EnrichPackageResourcesAsync (shared by both tiers)
• BuildRoleExcludeSet — extracted from duplicated logic in both Build methods
2. Refactored ConnectionQuery.cs — Now inherits ConnectionQueryCore:
• Keeps GetConnectionsAsync, GetConnectionsFromOthersAsync, GetConnectionsToOthersAsync
• Keeps enrichment-only methods: EnrichEntities, LoadResourcesByKeyAsync, LoadInstancesByKeyAsync
• Removed mutable _rightholderAssignments field — GetRightholderAssignments is now static
• ConnectionIndex<T> changed from internal to public
3. Registered ConnectionQueryCore in DI alongside ConnectionQuery

What this enables going forward
• PIP and AuthorizedParties consumers can be migrated to inject ConnectionQueryCore directly, calling GetBaseConnectionsAsync + LoadPackagesByKeyAsync without any entity enrichment overhead
• The Connection API continues using ConnectionQuery with full enrichment as before
• All 530 existing tests pass unchanged
